+++ sn_networking/CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.40](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.8.39...sn_networking-v0.8.40) - 2023-10-22
+
+### Added
+- *(protocol)* Nodes can error StoreCosts if they have data.
+
+## [0.8.39](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.8.38...sn_networking-v0.8.39) - 2023-10-21
+
+### Fixed
+- *(network)* return references when sorting peers
+- *(network)* prevent cloning of all our peers while sorting them
+
+## [0.8.38](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.8.37...sn_networking-v0.8.38) - 2023-10-20
+
+### Added
+- log network address with KBucketKey
+
 ## [0.8.37](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.8.36...sn_networking-v0.8.37) - 2023-10-20
 
 ### Added