+++ sn_node/CHANGELOG.md
@@ -1767,6 +1767,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - *(node)* [**breaking**] use `NodeBuilder` to construct and run node
 - *(node)* remove random get_closest on inactivity
 
+## [0.95.1](https://github.com/maidsafe/safe_network/compare/sn_node-v0.95.0...sn_node-v0.95.1) - 2023-10-20
+
+### Added
+- log network address with KBucketKey
+
+## [0.95.2](https://github.com/maidsafe/safe_network/compare/sn_node-v0.95.1...sn_node-v0.95.2) - 2023-10-21
+
+### Fixed
+- *(network)* return references when sorting peers
+- *(network)* prevent cloning of all our peers while sorting them
+
+## [0.95.3](https://github.com/maidsafe/safe_network/compare/sn_node-v0.95.2...sn_node-v0.95.3) - 2023-10-22
+
+### Added
+- *(protocol)* Nodes can error StoreCosts if they have data.
+
+### Other
+- *(node)* increase replication frequency
+
 ## v0.1.0 (2023-05-04)
 
 ### Chore